packages/skills/src/skills/design_automation/requirements.txt
# Design Automation Skill Dependencies

# Core dependencies
anthropic>=0.40.0
pydantic>=2.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Image processing
Pillow>=10.0.0
colorthief>=0.2.1

# Utilities
python-dotenv>=1.0.0